# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "rs-ibed",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.0.2"
  },
  "paths": {
    "/api/admin/images": {
      "get": {
        "tags": [
          "handlers::admin"
        ],
        "summary": "GET /api/admin/images",
        "operationId": "list_images",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number, starting from 1",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Fuzzy search by display_name",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "description": "Start date (YYYY-MM-DD)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "description": "End date (YYYY-MM-DD)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "deleted",
            "in": "query",
            "description": "Filter by deletion status; defaults to non-deleted images when omitted",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated image list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/images/count": {
      "get": {
        "tags": [
          "handlers::admin"
        ],
        "summary": "GET /api/admin/images/count",
        "operationId": "count_images",
        "responses": {
          "200": {
            "description": "Image counts grouped by deletion status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageCountResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/images/trash/empty": {
      "post": {
        "tags": [
          "handlers::admin"
        ],
        "summary": "POST /api/admin/images/trash/empty",
        "operationId": "empty_trash",
        "responses": {
          "200": {
            "description": "Trash emptied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/images/{id}": {
      "get": {
        "tags": [
          "handlers::admin"
        ],
        "summary": "GET /api/admin/images/:id",
        "operationId": "get_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Image ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/images/{id}/delete": {
      "post": {
        "tags": [
          "handlers::admin"
        ],
        "summary": "POST /api/admin/images/:id/delete",
        "operationId": "delete_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Image ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Soft deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/images/{id}/permanent-delete": {
      "post": {
        "tags": [
          "handlers::admin"
        ],
        "summary": "POST /api/admin/images/{id}/permanent-delete",
        "operationId": "permanent_delete_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Image ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permanently deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Image not in trash"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/images/{id}/rename": {
      "post": {
        "tags": [
          "handlers::admin"
        ],
        "summary": "POST /api/admin/images/:id/rename",
        "operationId": "rename_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Image ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Renamed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/images/{id}/restore": {
      "post": {
        "tags": [
          "handlers::admin"
        ],
        "summary": "POST /api/admin/images/:id/restore",
        "operationId": "restore_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Image ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Restored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/storage/endpoints": {
      "get": {
        "tags": [
          "handlers::admin"
        ],
        "summary": "GET /api/admin/storage/endpoints",
        "operationId": "list_storage_endpoints",
        "responses": {
          "200": {
            "description": "List of storage endpoints",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StorageEndpointResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/storage/endpoints/{name}/update": {
      "post": {
        "tags": [
          "handlers::admin"
        ],
        "summary": "POST /api/admin/storage/endpoints/{name}/update",
        "operationId": "update_storage_endpoint",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Endpoint name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEndpointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Cannot change priority or capacity via API"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DirectMode": {
        "type": "string",
        "description": "How clients access images stored on an S3 endpoint.",
        "enum": [
          "proxy",
          "presigned",
          "public"
        ]
      },
      "ImageCountResponse": {
        "type": "object",
        "required": [
          "active",
          "in_trash"
        ],
        "properties": {
          "active": {
            "type": "integer",
            "format": "int64"
          },
          "in_trash": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ImageDetailResponse": {
        "type": "object",
        "required": [
          "image",
          "view_url"
        ],
        "properties": {
          "direct_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "image": {
            "$ref": "#/components/schemas/ImageModel"
          },
          "storage_available": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Whether the image file is available on at least one active storage node.\nOnly populated in admin API responses; null means not checked."
          },
          "view_url": {
            "type": "string"
          }
        }
      },
      "ImageListItem": {
        "type": "object",
        "required": [
          "id",
          "hash",
          "display_name",
          "file_name",
          "extension",
          "mime_type",
          "size",
          "width",
          "height",
          "is_deleted",
          "created_at",
          "updated_at",
          "view_url"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "direct_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "display_name": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "file_name": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "is_deleted": {
            "type": "boolean"
          },
          "mime_type": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "storage_available": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Whether the image file is available on at least one active storage node.\nOnly populated in admin API responses; null means not checked."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "view_url": {
            "type": "string"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ImageListResponse": {
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "per_page"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageListItem"
            }
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "per_page": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ImageModel": {
        "type": "object",
        "required": [
          "id",
          "hash",
          "display_name",
          "file_name",
          "extension",
          "mime_type",
          "size",
          "width",
          "height",
          "is_deleted",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_name": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "file_name": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "is_deleted": {
            "type": "boolean"
          },
          "mime_type": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "width": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "OkResponse": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
      "RenameRequest": {
        "type": "object",
        "required": [
          "display_name"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          }
        }
      },
      "StorageEndpointResponse": {
        "type": "object",
        "description": "Response returned by GET /api/admin/storage/endpoints —\nmerges live in-memory state with config-immutable fields.",
        "required": [
          "name",
          "description",
          "endpoint_type",
          "capacity_bytes",
          "used_size",
          "priority",
          "status",
          "direct_mode"
        ],
        "properties": {
          "capacity_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string"
          },
          "direct_mode": {
            "$ref": "#/components/schemas/DirectMode",
            "description": "How clients access images: \"proxy\", \"presigned\", or \"public\"."
          },
          "endpoint_type": {
            "type": "string",
            "description": "Backend type from config: \"Local\" or \"S3\"."
          },
          "name": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "used_size": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "UpdateEndpointRequest": {
        "type": "object",
        "description": "Only `status` and `description` are mutable via the API.\n`priority` and `capacity_bytes` are config-immutable.",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "AUTH_TOKEN"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ibed_token"
      }
    }
  }
}